# Significantly improve render performance