This update represents a significant performance improvement for viewport aware rendering

```text
BenchmarkRenderTreeWithViewport_Wide/legacy_wide_1000-8                     1009           1272783 ns/op          472932 B/op       4101 allocs/op
BenchmarkRenderTreeWithViewport_Wide/legacy_wide_5000-8                      195           5916128 ns/op         1766394 B/op      20112 allocs/op
BenchmarkRenderTreeWithViewport_Wide/legacy_wide_10000-8                      87          12064861 ns/op         4191881 B/op      40121 allocs/op
BenchmarkRenderTreeWithViewport_Deep/legacy_d5_b3-8                         1860            615538 ns/op          319196 B/op       2121 allocs/op
BenchmarkRenderTreeWithViewport_Deep/legacy_d6_b3-8                          682           1695663 ns/op          690439 B/op       6499 allocs/op

BenchmarkRenderTreeWithViewport_Wide/1000-8                17308             67208 ns/op           81833 B/op        147 allocs/op
BenchmarkRenderTreeWithViewport_Wide/5000-8                 3894            299847 ns/op          602919 B/op        153 allocs/op
BenchmarkRenderTreeWithViewport_Wide/10000-8                1854            650096 ns/op         1157156 B/op        156 allocs/op
BenchmarkRenderTreeWithViewport_Deep/d5_b3-8               29205             41062 ns/op           18222 B/op        189 allocs/op
BenchmarkRenderTreeWithViewport_Deep/d6_b3-8               17007             70356 ns/op           31770 B/op        195 allocs/op
```